Move test to `test_overrides_parser.py`
